+++ predicators/ground_truth_models/touch_point/__init__.py
@@ -1,0 +1,5 @@
+"""Ground-truth models for touch point environment and variants."""
+
+from .nsrts import TouchPointGroundTruthNSRTFactory
+
+__all__ = ["TouchPointGroundTruthNSRTFactory"]
